Fix encoding of the suffix of CJK counter styles in UA style sheet
Many CJK counter styles use "、" (U+3001) as the suffix. This patch
properly escapes it in the UA style sheet to make the encoding correct.

As verification, this patch also adds ref files to the related WPT
tests. The new ref files are mechanically generated using the same
method as crrev.com/c/2674491

Bug: 687225
Change-Id: I4f29f17c290e21906601d8366a6a7db7d67b02a5